requirements.txt
pytesseract
Pillow
pyperclip
pynput
plyer
pystray
